requirements.txt
celery==4.1.0
coreapi==2.3.3
Django==2.0.3
djangorestframework==3.7.7
